CVMaker(Pending)/archieve/requirements.txt
flask==3.0.0
pdfplumber==0.10.3
weasyprint==60.1
langdetect==1.0.9
jinja2==3.1.2 
requests==2.31.0
python-dotenv==1.0.0